runner: expand result size to u64

+++ src/boilerplate.zig
@@ -1,5 +1,5 @@
 const std = @import("std");
 
-pub const AoCResult = struct { part1: u32, part2: u32 };
+pub const AoCResult = struct { part1: u64, part2: u64 };
 
 pub const Solver = fn (file: std.fs.File) anyerror!AoCResult;
